+++ libopendmi/test/bswap64.c
@@ -36,7 +36,7 @@ int main(void)
 
 static bool test_bswap(uint64_t (*func)(uint64_t))
 {
-    for (unsigned i = 0; i < dmi_array_size(test_data); i++) {
+    for (size_t i = 0; i < dmi_array_size(test_data); i++) {
         if (func(test_data[i].value) != test_data[i].result)
             return false;
     }
