+++ libopendmi/test/decode-bcd.c
@@ -66,7 +66,7 @@ struct test_vector test_data[] =
 
 int main(void)
 {
-    for (unsigned i = 0; i < dmi_array_size(test_data); i++) {
+    for (size_t i = 0; i < dmi_array_size(test_data); i++) {
         if (dmi_decode_bcd(test_data[i].value, test_data[i].length) != test_data[i].result)
             return EXIT_FAILURE;
     }
