+++ libopendmi/test/memory-device-size.c
@@ -11,7 +11,7 @@
 
 int main(void)
 {
-    for (unsigned i = 0; i < 15; i++) {
+    for (size_t i = 0; i < 15; i++) {
         if (dmi_memory_device_size(1u << i) != (dmi_size_t)(1u << i) * 1048576)
             return EXIT_FAILURE;
