+++ opendmi/src/format/text/handlers.c
@@ -234,7 +234,7 @@ void dmi_text_entity_attr_set(
     else
         return;
 
-    for (unsigned i = 0; i < attr->size * CHAR_BIT; i++) {
+    for (size_t i = 0; i < attr->size * CHAR_BIT; i++) {
         const char *name = dmi_name_lookup(attr->params.values, i);
         if (!name)
             continue;
